backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.1
langgraph==0.0.24
langchain-community==0.0.13
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1
duckduckgo-search
